Fix DMA to/from ROML/ROMH/IO1/IO2 areas
If the C64 drives ROML, ROMH, IO1 or IO2 during a DMA cycle, let the
DMA sequencer continue to drive the address bus, and disable the data
levelshifter and drive the data bus locally.  This way the correct
values will be seen on the address and data bus both by the DMA engine
and by the IO register implementations (where applicable), while the
C64 will continue to drive the appropriate select signal.

Also, make sure to not load d_q_reg with ioefdata during a write
cycle, since that would overwrite the value that the DMA engine is
trying to write.